+++ deltafi-api/helpers/helpers.rb
@@ -57,3 +57,9 @@ class ApiServer < Sinatra::Base
     end
   end
 end
+
+class Sinatra::Helpers::Stream
+  def send_heartbeat
+    self << "event: heartbeat\ndata: #{Time.now.to_i}\n\n"
+  end
+end